v2/approaches/docker_packaging/requirements.txt
torch==1.10.1
torchvision==0.11.1
Pillow==8.4.0
